# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autonomous Task & Data Orchestration Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelemetrySnapshot {
  cpuPercent: number;
  memoryPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  uptimeSeconds: number;
  requestsPerSecond: number;
  activeConnections: number;
  timestamp: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  queued: 'queued',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type TaskAgent = typeof TaskAgent[keyof typeof TaskAgent];


export const TaskAgent = {
  claude: 'claude',
  deepseek: 'deepseek',
  hermes: 'hermes',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Task {
  id: string;
  payload: string;
  status: TaskStatus;
  agent: TaskAgent;
  priority: TaskPriority;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputAgent = typeof TaskInputAgent[keyof typeof TaskInputAgent];


export const TaskInputAgent = {
  claude: 'claude',
  deepseek: 'deepseek',
  hermes: 'hermes',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  payload: string;
  agent?: TaskInputAgent;
  priority?: TaskInputPriority;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  idle: 'idle',
  processing: 'processing',
  error: 'error',
  offline: 'offline',
} as const;

export interface Agent {
  id: string;
  name: string;
  status: AgentStatus;
  tasksProcessed: number;
  avgLatencyMs: number;
  model: string;
  /** @nullable */
  currentTaskId?: string | null;
}

export interface StatsSummary {
  totalTasks: number;
  completedTasks: number;
  failedTasks: number;
  queuedTasks: number;
  processingTasks: number;
  avgThroughput: number;
  topAgent: string;
}

export interface ErrorResponse {
  error: string;
}