# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autonomous Task & Data Orchestration Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStatus } from './agentStatus';

export interface Agent {
  id: string;
  name: string;
  status: AgentStatus;
  tasksProcessed: number;
  avgLatencyMs: number;
  model: string;
  /** @nullable */
  currentTaskId?: string | null;
}