# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autonomous Task & Data Orchestration Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalTasks: number;
  completedTasks: number;
  failedTasks: number;
  queuedTasks: number;
  processingTasks: number;
  avgThroughput: number;
  topAgent: string;
}