# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autonomous Task & Data Orchestration Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputAgent } from './taskInputAgent';
import type { TaskInputPriority } from './taskInputPriority';

export interface TaskInput {
  /** @minLength 1 */
  payload: string;
  agent?: TaskInputAgent;
  priority?: TaskInputPriority;
}