# lib/api-zod/src/generated/types/telemetrySnapshot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autonomous Task & Data Orchestration Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TelemetrySnapshot {
  cpuPercent: number;
  memoryPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  uptimeSeconds: number;
  requestsPerSecond: number;
  activeConnections: number;
  timestamp: string;
}